+++ src/main/java/com/example/mohago_nocar/plan/application/v2/PlanTravelCourseResponseDtoV2.java
@@ -1,0 +1,9 @@
+package com.example.mohago_nocar.plan.application.v2;
+
+import lombok.Builder;
+
+@Builder
+public record PlanTravelCourseResponseDtoV2(
+        String batch_id
+) {
+}
