--- src/main/java/com/example/mohago_nocar/plan/presentation/TravelPlanResponseDtoV2.java
@@ -1,6 +1,0 @@
-package com.example.mohago_nocar.plan.presentation;
-
-public record TravelPlanResponseDtoV2(
-        String batch_id
-) {
-}
